requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
DateTime==5.4
distro==1.9.0
dnspython==2.5.0
exceptiongroup==1.2.0
fastapi==0.109.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.6
langchain-community==0.0.20
langchain-core==0.1.22
langchain-openai==0.0.5
langsmith==0.0.87
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.12.0
packaging==23.2
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.6.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.26
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==1.26.6
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
zope.interface==6.1
